requirements.txt
pandas==0.23.4
jieba==0.42.1
easydict==1.9
numpy==1.15.1
torch==1.4.0+cpu
gensim==3.8.3
tqdm==4.26.0
PyYAML==5.3.1
scikit_learn==0.23.2
